Added ProductService.cs validations

+++ SuperMarketWebApi.Application/Services/ProductService.cs
@@ -1,6 +1,7 @@
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SuperMarketWebApi.Core.Entities;
+using SuperMarketWebApi.Core.Exceptions;
 using SuperMarketWebApi.Core.Records;
 using SuperMarketWebApi.Persistence.Contexts;
 
@@ -26,12 +27,35 @@ public class ProductService : IProductService
         var productById = await _context.ProductInfo
             .Where(p => p.Id == id)
             .FirstOrDefaultAsync(ct);
-            
+
+        if (productById is null)
+        {
+            throw new ProductException("Can not find product with this id");
+        }
+        
         return productById;
     }
-    
+
     public async Task CreateNewProduct(CreateNewProductRequest request, CancellationToken ct)
     {
+        if (string.IsNullOrWhiteSpace(request.Name.Trim()))
+        {
+            throw new ProductException("Can not create product with blank name");
+        }
+
+        if (request.Price <= 0)
+        {
+            throw new ProductException("Can not create product with negative or zero price");
+        }
+
+        var categories = await _context.ProductCategory
+            .ToListAsync(ct);
+
+        if (!categories.Any(pc => pc.Id == request.CategoryId))
+        {
+            throw new ProductException("This category does not exist");
+        }
+
         var newProduct = new ProductInfo()
         {
             Name = request.Name,
@@ -40,16 +64,31 @@ public class ProductService : IProductService
             Photo = request.Photo,
             Price = request.Price
         };
-
+        
         await _context.ProductInfo.AddAsync(newProduct, ct);
         await _context.SaveChangesAsync(ct);
     }
     
     public async Task UpdateProductById(int id, UpdateProductRequest request, CancellationToken ct)
     {
+        if (string.IsNullOrWhiteSpace(request.Name.Trim()))
+        {
+            throw new ProductException("Can not create product with blank name");
+        }
+
+        if (request.Price <= 0)
+        {
+            throw new ProductException("Can not create product with negative or zero price");
+        }
+        
         var productToUpdate = await _context.ProductInfo
             .Where(p => p.Id == id)
             .FirstOrDefaultAsync(ct);
+
+        if (productToUpdate is null)
+        {
+            throw new ProductException("Can not find product with this id");
+        }
 
         productToUpdate.Name = request.Name;
         productToUpdate.Description = request.Description;
@@ -65,6 +104,11 @@ public class ProductService : IProductService
         var productToDelete = await _context.ProductInfo
             .Where(p => p.Id == id)
             .FirstOrDefaultAsync(ct);
+        
+        if (productToDelete is null)
+        {
+            throw new ProductException("Can not find product with this id");
+        }
 
         _context.ProductInfo.Remove(productToDelete);
         await _context.SaveChangesAsync(ct);
